+++ internal/lido/contracts/csmodule/keys.go
@@ -48,14 +48,15 @@ func KeysStatus(network string, nodeID *big.Int) (Keys, error) {
 	if nodeID.Sign() < 0 {
 		return keys, fmt.Errorf("node ID value out-of-bounds: can't be negative")
 	}
-	contract, err := csModuleContract(network)
+	contract, client, err := csModuleContract(network)
 	if err != nil {
 		return keys, fmt.Errorf("failed to call csModuleContract: %w", err)
 	}
+	defer client.Close()
 
 	nodeOp, err := contract.GetNodeOperatorSummary(nil, nodeID)
 	if err != nil {
-		return keys, fmt.Errorf("failed to call GetNodeOperator: %w", err)
+		return keys, fmt.Errorf("failed to call GetNodeOperatorSummary contract method: %w", err)
 	}
 	keys.DepositableValidatorsCount = nodeOp.DepositableValidatorsCount
 	keys.DepositedValidators = nodeOp.TotalDepositedValidators